refactor: inline startDevServer fn
